+++ src/shared/components/query/QueryConnector.ts
