+++ src/shared/components/query/styles.module.scss
